Split the integration test into e2e dirs
