+++ go.mod
@@ -29,7 +29,6 @@ require (
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp v1.35.0
 	go.opentelemetry.io/otel/sdk v1.35.0
 	go.opentelemetry.io/otel/trace v1.35.0
-	go.uber.org/goleak v1.3.0
 	golang.org/x/arch v0.16.0
 	golang.org/x/sys v0.32.0
 	google.golang.org/grpc v1.71.1
