+++ internal/test/e2e/kafka-go/cmd/dependencies.Dockerfile
