+++ internal/test/e2e/otelglobal/cmd/main.go
